requirements.txt
fastapi>=0.65.2  # version fixed due to complete re-overriding fastapi objects
uvicorn
asyncpg
passlib
argon2-cffi  # required for passlib.hash.argon2
pyjwt
databases~=0.4.3
python-dotenv~=0.18.0